New major version with improved UI.

+++ app/src/test/java/com/panosdim/moneytrack/ExampleUnitTest.kt
@@ -1,17 +1,16 @@
 package com.panosdim.moneytrack
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
  *
- * @see [Testing documentation](http://d.android.com/tools/testing)
+ * See [testing documentation](http://d.android.com/tools/testing).
  */
 class ExampleUnitTest {
     @Test
     fun addition_isCorrect() {
-        assertEquals(4, (2 + 2).toLong())
+        assertEquals(4, 2 + 2)
     }
 }
